+++ src/lib/lib.rs
@@ -1,6 +1,4 @@
 pub mod board;
-pub mod board_logic;
 pub mod errors;
-pub mod evaluator;
 pub mod file_reader;
 pub mod util;
